home_app_proj/requirements.txt
django-phonenumber-field==5.0.0
phonenumbers==8.12.20